Added more functions to Assembly and some in header.h
Added hexToBinary converter of one byte in header.
Added getOP function that will give us the opcode from a byte.
Corrected getFormatType, now it returns 1,2,3, or 4 depending on three nibbles.
The addFormat3 and addFormat4 functions are functional but are just for debuging use since it does not calculate the operand addresses and the labels yet.